got rid of type File  proceed with SingleFile with option routemask true

+++ lib/Ingestor.py
@@ -17,7 +17,6 @@ named COPYING in the root of the source directory tree.
 # Description:
 #
 # Revision :  2006-05-14 (ingestCollection by MG)
-#             2006-10-14 (ingestFile       by MG)
 #
 #############################################################################################
 
@@ -58,15 +57,16 @@ class Ingestor(object):
         self.pxManager.initNames()                # Set rx and tx names
         self.clientNames = self.pxManager.getTxNames()         # Obtains the list of client's names (the ones to wich we can link files)
         self.filterNames = self.pxManager.getFxNames()         # Obtains the list of filter's names (the ones to wich we can link files)
+        if source is not None:
+           if self.source.name in self.filterNames : self.filterNames.remove(self.source.name)
         self.sourlientNames = self.pxManager.getTRxNames()     # Obtains the list of sourlient's names (the ones to wich we can link files)
         self.allNames = self.clientNames + self.filterNames + self.sourlientNames # Clients + Sourlients names
         self.clients = {}   # All the Client/Filter/Sourlient objects
         self.dbDirsCache = CacheManager(maxEntries=200000, timeout=25*3600)      # Directories created in the DB
         self.clientDirsCache = CacheManager(maxEntries=100000, timeout=2*3600)   # Directories created in TXQ
         self.feedNames = []  # source to feed
-        self.feeds = {}  # source to feed
+        self.feeds = {}
         if source is not None:
-           if self.source.name in self.allNames : self.allNames.remove(self.source.name)
            self.logger.info("Ingestor (source %s) can link files to clients: %s" % (source.name, self.allNames))
 
     def setFeeds(self, feedNames ):
@@ -99,7 +99,7 @@ class Ingestor(object):
         from Sourlient import Sourlient
         for name in self.clientNames:
             self.clients[name] = Client(name, self.logger)
-        for name in self.filterNames:
+        for name in self.filterNames :
             self.clients[name] = Source(name, self.logger, False, True)
         for name in self.sourlientNames:
             self.clients[name] = Sourlient(name, self.logger, False)
@@ -188,6 +188,7 @@ class Ingestor(object):
         """
         Verify if ingestName is matching one mask of a client
         """
+        from Source import Source
 
         if len(client.masks_deprecated) > 0 :
            for mask in client.masks_deprecated:
@@ -205,23 +206,12 @@ class Ingestor(object):
                if len(mask) == 3 : return True
                return False
 
+        if isinstance(client,Source) : return True
+
         return False
-
-    def getMatchingClientNamesFromKey(self, key, ingestName):
-
-        matchingClients = self.drp.getClients(key)
-
-        # check pattern matching in clients
-
-        if self.source.clientsPatternMatching:
-           matchingClients = self.getMatchingClientNamesFromMasks(ingestName, matchingClients)
-                
-        return matchingClients
 
     def getMatchingClientNamesFromMasks(self, ingestName, potentialClientNames):
         matchingClientNames = []
-
-        if potentialClientNames == None : return None
 
         for name in potentialClientNames:
             try:
@@ -242,15 +232,13 @@ class Ingestor(object):
                 pass
         return matchingFeedNames
 
-    def ingest(self, receptionName, ingestName, clientNames, header=None ):
+    def ingest(self, receptionName, ingestName, clientNames ):
         self.logger.debug("Reception Name: %s" % receptionName)
         dbName = self.getDBName(ingestName)
 
         if dbName == '':
             self.logger.warning('Bad ingest name (%s) => No dbName' % ingestName)
             return 0
-
-        if clientNames == None : clientNames = []
         
         self.createDir(os.path.dirname(dbName), self.dbDirsCache)
         #try:
@@ -288,13 +276,8 @@ class Ingestor(object):
         if ingestName.find("PROBLEM_BULLETIN") is not -1:
             return 1
 
-        priority=None
-        if header != None :
-           key      = self.drp.getKeyFromHeader(header)
-           priority = self.drp.getHeaderPriority(key)
-
         for name in clientNames:
-            clientQueueName = self.getClientQueueName(name, ingestName, priority )
+            clientQueueName = self.getClientQueueName(name, ingestName)
             self.createDir(os.path.dirname(clientQueueName), self.clientDirsCache)
             #try:
             #    os.link(dbName, clientQueueName)
@@ -320,104 +303,18 @@ class Ingestor(object):
     def run(self):
         if self.source.type == 'single-file':
             self.ingestSingleFile()
-        elif self.source.type == 'file':
-            self.ingestFile()
         elif self.source.type == 'bulletin-file':
             self.ingestBulletinFile()
         elif self.source.type == 'collector':
             self.ingestCollection()
 
-    def ingestFile(self, igniter):
-        from DiskReader import DiskReader
-
-        self.drp = DirectRoutingParser(PXPaths.ROUTING_TABLE, self.allNames, self.logger)
-        self.drp.parse()
-
-        # we will not use paternMaching ... file validation now require
-        # building a key out of the filename and match the key with the routingtable
-        patternMatching = False
-
-        reader = DiskReader(self.ingestDir, self.source.batch, self.source.validation, patternMatching,
-                            self.source.mtime, False, self.source.logger, self.source.sorter, self.source)
-
-        while True:
-            if igniter.reloadMode == True:
-                # We assign the defaults, reread configuration file for the source
-                # and reread all configuration file for the clients (all this in __init__)
-                self.source.__init__(self.source.name, self.source.logger)
-
-
-                self.drp = DirectRoutingParser(PXPaths.ROUTING_TABLE, self.allNames, self.logger)
-                self.drp.parse()
-
-                reader = DiskReader(self.ingestDir, self.source.batch, self.source.validation, patternMatching,
-                                    self.source.mtime, False, self.source.logger, self.source.sorter, self.source)
-                self.logger.info("Receiver has been reloaded")
-                igniter.reloadMode = False
-            reader.read()
-            if len(reader.sortedFiles) <= 0:
-               time.sleep(1)
-               continue
-
-            sortedFiles = reader.sortedFiles[:self.source.batch]
-            self.logger.info("%d files will be ingested" % len(sortedFiles))
-
-            # loop on all files
-
-            for filex in sortedFiles:
-                file = filex
-
-                # converting the file if necessary
-
-                if self.source.execfile != None :
-                   fxfile = self.source.run_fx_script(file,self.source.logger)
-
-                   # convertion did not work
-                   if fxfile == None :
-                          self.logger.warning("FX script ignored the file : %s"    % os.path.basename(file) )
-                          os.unlink(file)
-                          continue
-
-                   # file already in proper format
-                   elif fxfile == file :
-                          self.logger.warning("FX script kept the file as is : %s" % os.path.basename(file) )
-
-                   # file converted...
-                   else :
-                          self.logger.info("FX script modified %s to %s " % (os.path.basename(file),fxfile) )
-                          os.unlink(file)
-                          file = fxfile
-
-                # ingestName is used for key generation
-                # Key is used to get clientlist
-
-                ingestName       = self.getIngestName(os.path.basename(file))
-                routeKey         = self.getRouteKey(ingestName)
-                self.logger.debug("routeKey = %s" % routeKey )
-
-                # if key is found : find clients associated with key
-                # else ... file is ingested as problem....
-
-                matchingClients = []
-
-                if routeKey != None :
-                   matchingClients = self.getMatchingClientNamesFromKey(routeKey,ingestName)
-
-                if routeKey == None or matchingClients == None :
-                   self.count = self.count + 1
-                   if self.count > self.Mcount : self.count = 0
-                   strCount = string.zfill(self.count, len(str(self.Mcount)))
-                   parts=ingestName.split(':')
-                   ingestName =  parts[0] + "_" + strCount + ":PROBLEM:PROBLEM:PROBLEM:PROBLEM:PROBLEM"
-                   matchingClients = []
-
-                # ingesting the file
-
-                self.ingest(file, ingestName, matchingClients, routeKey  )
-                os.unlink(file)
-
     def ingestSingleFile(self, igniter):
         from DiskReader import DiskReader
+        from DirectRoutingParser import DirectRoutingParser
+
+        if self.source.routemask :
+           self.drp = DirectRoutingParser(self.source.routingTable, self.allNames, self.logger, self.source.routing_version)
+           self.drp.parse()
 
         reader = DiskReader(self.ingestDir, self.source.batch, self.source.validation, self.source.patternMatching,
                             self.source.mtime, False, self.source.logger, self.source.sorter, self.source)
@@ -427,6 +324,10 @@ class Ingestor(object):
                 # We assign the defaults, reread configuration file for the source
                 # and reread all configuration file for the clients (all this in __init__)
                 self.source.__init__(self.source.name, self.source.logger)
+
+                if self.source.routemask :
+                   self.drp = DirectRoutingParser(self.source.routingTable, self.allNames, self.logger)
+                   self.drp.parse()
 
                 reader = DiskReader(self.ingestDir, self.source.batch, self.source.validation, self.source.patternMatching,
                                     self.source.mtime, False, self.source.logger, self.source.sorter, self.source)
@@ -464,10 +365,25 @@ class Ingestor(object):
                           os.unlink(file)
                           file = fxfile
 
-                # ingesting the file
+                # filename to ingest
 
                 ingestName = self.getIngestName(os.path.basename(file))
-                matchingClients = self.getMatchingClientNamesFromMasks(ingestName, self.clientNames)
+
+                # usual clients
+
+                potentials = self.clientNames + self.filterNames
+
+                # routing clients (accept mask)
+
+		if self.source.routemask :
+                   potentials = []
+                   key = self.getRouteKey(ingestName)
+		   if key != None :
+		      lst = self.drp.getClients(key)
+		      if lst != None : potentials = lst
+
+                # ingesting the file
+                matchingClients  = self.getMatchingClientNamesFromMasks(ingestName, potentials )
                 self.logger.debug("Matching (from patterns) client names: %s" % matchingClients)
                 self.ingest(file, ingestName, matchingClients )
                 os.unlink(file)
